refactor: extract buildImageResponse to a native-free module (Codex Round 3 P2)
The screenshot-response unit test imported buildImageResponse from
screenshot.ts, which transitively loads engine/nutjs → native libnut. On a
host without the native deps (e.g. a Linux unit lane) that import aborts with
"libXtst.so.6: cannot open shared object file", so the cache-response tests
never ran there (Codex reproduced it).

Move buildImageResponse into src/tools/screenshot-response.ts, which depends
only on the pure cache layer (screenshot-cache) + tool result types — no
native bindings. screenshot.ts re-imports it; the test imports the native-free
module. Behaviour unchanged.

Tests: screenshot-response 5 green, tool-naming 133 green. tsc + lint clean.
